Guard against `configure` run before bootstrap
If `configure` is run before `bootstrap` abort configure and
tell the user bootstrap is needed. This is mostly likely to
be an issue after a `reset` as it may not be clear to the
user that a boostrap is needed.

(cherry picked from commit 308dcf359d18bd58c6924b606f557eb63fd89f73)